Added dedicated astro integration and fixed issue with ending write stream while also calling next middleware
