+++ packages/libs/index.ts
@@ -1,21 +1,9 @@
-import {ViteDevServer} from "vite";
 import {ServerMiddleWare} from "./server";
-import {buildMiddleWare} from "./build";
-import {IConfig} from "./types";
-
-function DynamicPublicDirectory(assets: string[], opts: IConfig = {}): any {
-    return {
-        configureServer(server: ViteDevServer) {
-            return ServerMiddleWare({server, assets, options: opts})
-        },
-        async writeBundle(options) {
-            buildMiddleWare(options, assets)
-        },
-        name: "dynamic assets"
-    };
-}
+import {DynamicPublicDirectory} from "./dynamicPublicDirectory";
+import {AstroIntegration} from "./astroIntegration";
 
 export {
     DynamicPublicDirectory,
-    ServerMiddleWare
+    ServerMiddleWare,
+    AstroIntegration as astroMultipleAssets
 }
